Add new flush consolidation
This version largely respects the flush requests from the remote server.
In this sense, it is quite conservative and must be paired with a server
that does sensible flushing.

This currently does not yet completely work: the packets still appear to
be split up for me, even though they shouldn't be. I'll have to investigate
this later.